refactor: guard for non-specified filters + move checkPorts to utils
